requirements.txt
ctransformers==0.2.27
sentence-transformers==2.3.1 #For creating sentence embeddings
langchain-pinecone
langchain-community
langchain==0.1.8
python-dotenv==1.0.1
flask==3.0.2
pypdf

-e . 